Move deprecated panel classes to wagtail.admin.panels.deprecated
